Defined and Updated Controllers

+++ app/Http/Controllers/PostController.php
@@ -3,8 +3,13 @@
 namespace App\Http\Controllers;
 
 use App\BlogPost;
+use App\Image;
+use App\Events\BlogPostPosted;
+use App\Facades\CounterFacade;
 use App\Http\Requests\StorePost;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Cache;
+use Illuminate\Support\Facades\Storage;
 
 class PostController extends Controller
 {
@@ -15,12 +20,17 @@ class PostController extends Controller
 
     public function index()
     {
-        return view('posts.index', ['posts' => BlogPost::withCount('comments')->get()]);
+        return view('posts.index', ['posts' => BlogPost::latestWithRelations()->get()]);
     }
 
     public function show($id)
     {
-        return view('posts.show', ['post' => BlogPost::with('comments')->findOrFail($id)]);
+        $blogPost = Cache::tags(['blog-post'])->remember("blog-post-{$id}", 60, function() use ($id){
+            return BlogPost::with('comments', 'tags', 'user', 'comments.user')->findOrFail($id);
+        });
+        return view('posts.show', ['post' => $blogPost, 'counter' => CounterFacade::increment("blog-post-{$id}", ['blog-post']),
+
+        ]);
     }
 
     public function create()
@@ -31,7 +41,15 @@ class PostController extends Controller
     public function store(StorePost $request)
     {
         $validatedData = $request->validated();
+        $validatedData['user_id'] = $request->user()->id;
         $blogPost = BlogPost::create($validatedData);
+
+        if($request->hasFile('thumbnail')){
+            $path = $request->file('thumbnail')->store('thumbnails');
+            $blogPost->image()->save(Image::make(['path' => $path]));
+        }
+
+        event(new BlogPostPosted($blogPost));
         $request->session()->flash('status', 'Blog post was created!');
         return redirect()->route('posts.show', ['post' => $blogPost->id]);
     }
@@ -39,24 +57,45 @@ class PostController extends Controller
     public function edit($id)
     {
         $post = BlogPost::findOrFail($id);
+        $this->authorize($post);
+
         return view('posts.edit', ['post' => $post]);
     }
 
     public function update(StorePost $request, $id)
     {
         $post = BlogPost::findOrFail($id);
+        $this->authorize($post);
+
         $validatedData = $request->validated();
+
         $post->fill($validatedData);
+
+        if($request->hasFile('thumbnail')){
+            $path = $request->file('thumbnail')->store('thumbnails');
+
+            if($post->image){
+                Storage::delete($post->image->path);
+                $post->image->path = $path;
+                $post->image->save();
+            }else{
+                $post->image()->save(
+                    Image::make(['path' => $path])
+                );
+            }
+        }
+
         $post->save();
-        $request->session()->flash('status', 'BlogPost was updated!');
+        $request->session()->flash('status', 'Blog post was updated');
+
         return redirect()->route('posts.show', ['post' => $post->id]);
     }
 
     public function destroy(Request $request, $id)
     {
         $post = BlogPost::findOrFail($id);
+        $this->authorize($post);
         $post->delete();
-
         $request->session()->flash('status', 'Blog post was deleted!');
         return redirect()->route('posts.index');
     }
